Move message classes to s2c/c2c packages
