+++ src/main/java/org/saltyrtc/client/messages/c2c/Key.java
@@ -6,30 +6,45 @@
  * copied, modified, or distributed except according to those terms.
  */
 
-package org.saltyrtc.client.messages;
+package org.saltyrtc.client.messages.c2c;
+
+import com.neilalexander.jnacl.NaCl;
 
 import org.msgpack.core.MessagePacker;
 import org.saltyrtc.client.exceptions.ValidationError;
 import org.saltyrtc.client.helpers.ValidationHelper;
+import org.saltyrtc.client.messages.Message;
 
 import java.io.IOException;
 import java.util.Map;
 
-public class NewInitiator extends Message {
+public class Key extends Message {
 
-    public static String TYPE = "new-initiator";
+    public static String TYPE = "key";
 
-    public NewInitiator() { }
+    private byte[] key;
 
-    public NewInitiator(Map<String, Object> map) throws ValidationError {
+    public Key(byte[] key) {
+        this.key = key;
+    }
+
+    public Key(Map<String, Object> map) throws ValidationError {
         ValidationHelper.validateType(map.get("type"), TYPE);
+        this.key = ValidationHelper.validateByteArray(map.get("key"), NaCl.PUBLICKEYBYTES, "Key");
+    }
+
+    public byte[] getKey() {
+        return key;
     }
 
     @Override
     public void write(MessagePacker packer) throws IOException {
-        packer.packMapHeader(1)
+        packer.packMapHeader(2)
                 .packString("type")
-                    .packString(TYPE);
+                    .packString(TYPE)
+                .packString("key")
+                    .packBinaryHeader(this.key.length)
+                    .writePayload(this.key);
     }
 
     @Override
