+++ src/main/java/org/saltyrtc/client/messages/s2c/NewInitiator.java
@@ -6,44 +6,31 @@
  * copied, modified, or distributed except according to those terms.
  */
 
-package org.saltyrtc.client.messages;
-
-import com.neilalexander.jnacl.NaCl;
+package org.saltyrtc.client.messages.s2c;
 
 import org.msgpack.core.MessagePacker;
 import org.saltyrtc.client.exceptions.ValidationError;
 import org.saltyrtc.client.helpers.ValidationHelper;
+import org.saltyrtc.client.messages.Message;
 
 import java.io.IOException;
 import java.util.Map;
 
-public class ClientHello extends Message {
+public class NewInitiator extends Message {
 
-    public static String TYPE = "client-hello";
+    public static String TYPE = "new-initiator";
 
-    private byte[] key;
+    public NewInitiator() { }
 
-    public ClientHello(byte[] key) {
-        this.key = key;
-    }
-
-    public ClientHello(Map<String, Object> map) throws ValidationError {
+    public NewInitiator(Map<String, Object> map) throws ValidationError {
         ValidationHelper.validateType(map.get("type"), TYPE);
-        this.key = ValidationHelper.validateByteArray(map.get("key"), NaCl.PUBLICKEYBYTES, "Key");
-    }
-
-    public byte[] getKey() {
-        return key;
     }
 
     @Override
     public void write(MessagePacker packer) throws IOException {
-        packer.packMapHeader(2)
+        packer.packMapHeader(1)
                 .packString("type")
-                    .packString(TYPE)
-                .packString("key")
-                    .packBinaryHeader(this.key.length)
-                    .writePayload(this.key);
+                    .packString(TYPE);
     }
 
     @Override
