+++ src/main/java/org/saltyrtc/client/signaling/InitiatorSignaling.java
@@ -25,13 +25,13 @@ import org.saltyrtc.client.helpers.MessageReader;
 import org.saltyrtc.client.keystore.AuthToken;
 import org.saltyrtc.client.keystore.Box;
 import org.saltyrtc.client.keystore.KeyStore;
-import org.saltyrtc.client.messages.Auth;
-import org.saltyrtc.client.messages.DropResponder;
-import org.saltyrtc.client.messages.InitiatorServerAuth;
-import org.saltyrtc.client.messages.Key;
+import org.saltyrtc.client.messages.c2c.Auth;
+import org.saltyrtc.client.messages.s2c.DropResponder;
+import org.saltyrtc.client.messages.s2c.InitiatorServerAuth;
+import org.saltyrtc.client.messages.c2c.Key;
 import org.saltyrtc.client.messages.Message;
-import org.saltyrtc.client.messages.NewResponder;
-import org.saltyrtc.client.messages.Token;
+import org.saltyrtc.client.messages.s2c.NewResponder;
+import org.saltyrtc.client.messages.c2c.Token;
 import org.saltyrtc.client.nonce.CombinedSequence;
 import org.saltyrtc.client.nonce.SignalingChannelNonce;
 import org.saltyrtc.client.signaling.state.ResponderHandshakeState;
