+++ src/main/java/org/saltyrtc/client/signaling/ResponderSignaling.java
@@ -26,13 +26,13 @@ import org.saltyrtc.client.helpers.MessageReader;
 import org.saltyrtc.client.keystore.AuthToken;
 import org.saltyrtc.client.keystore.Box;
 import org.saltyrtc.client.keystore.KeyStore;
-import org.saltyrtc.client.messages.Auth;
-import org.saltyrtc.client.messages.ClientHello;
-import org.saltyrtc.client.messages.Key;
+import org.saltyrtc.client.messages.c2c.Auth;
+import org.saltyrtc.client.messages.s2c.ClientHello;
+import org.saltyrtc.client.messages.c2c.Key;
 import org.saltyrtc.client.messages.Message;
-import org.saltyrtc.client.messages.NewInitiator;
-import org.saltyrtc.client.messages.ResponderServerAuth;
-import org.saltyrtc.client.messages.Token;
+import org.saltyrtc.client.messages.s2c.NewInitiator;
+import org.saltyrtc.client.messages.s2c.ResponderServerAuth;
+import org.saltyrtc.client.messages.c2c.Token;
 import org.saltyrtc.client.nonce.CombinedSequence;
 import org.saltyrtc.client.nonce.SignalingChannelNonce;
 import org.saltyrtc.client.signaling.state.InitiatorHandshakeState;
