+++ src/main/java/org/saltyrtc/client/signaling/Signaling.java
@@ -34,14 +34,14 @@ import org.saltyrtc.client.helpers.MessageReader;
 import org.saltyrtc.client.keystore.AuthToken;
 import org.saltyrtc.client.keystore.Box;
 import org.saltyrtc.client.keystore.KeyStore;
-import org.saltyrtc.client.messages.Auth;
-import org.saltyrtc.client.messages.ClientAuth;
-import org.saltyrtc.client.messages.InitiatorServerAuth;
+import org.saltyrtc.client.messages.c2c.Auth;
+import org.saltyrtc.client.messages.s2c.ClientAuth;
+import org.saltyrtc.client.messages.s2c.InitiatorServerAuth;
 import org.saltyrtc.client.messages.Message;
-import org.saltyrtc.client.messages.ResponderServerAuth;
+import org.saltyrtc.client.messages.s2c.ResponderServerAuth;
 import org.saltyrtc.client.messages.Restart;
-import org.saltyrtc.client.messages.SendError;
-import org.saltyrtc.client.messages.ServerHello;
+import org.saltyrtc.client.messages.s2c.SendError;
+import org.saltyrtc.client.messages.s2c.ServerHello;
 import org.saltyrtc.client.nonce.CombinedSequence;
 import org.saltyrtc.client.nonce.CombinedSequencePair;
 import org.saltyrtc.client.nonce.DataChannelNonce;
