+++ src/test/java/org/saltyrtc/client/tests/messages/MessageTest.java
@@ -13,20 +13,20 @@ import org.saltyrtc.client.exceptions.SerializationError;
 import org.saltyrtc.client.exceptions.ValidationError;
 import org.saltyrtc.client.helpers.MessageReader;
 import org.saltyrtc.client.helpers.RandomHelper;
-import org.saltyrtc.client.messages.Auth;
-import org.saltyrtc.client.messages.ClientAuth;
-import org.saltyrtc.client.messages.ClientHello;
-import org.saltyrtc.client.messages.DropResponder;
-import org.saltyrtc.client.messages.InitiatorServerAuth;
-import org.saltyrtc.client.messages.Key;
+import org.saltyrtc.client.messages.c2c.Auth;
+import org.saltyrtc.client.messages.s2c.ClientAuth;
+import org.saltyrtc.client.messages.s2c.ClientHello;
+import org.saltyrtc.client.messages.s2c.DropResponder;
+import org.saltyrtc.client.messages.s2c.InitiatorServerAuth;
+import org.saltyrtc.client.messages.c2c.Key;
 import org.saltyrtc.client.messages.Message;
-import org.saltyrtc.client.messages.NewInitiator;
-import org.saltyrtc.client.messages.NewResponder;
-import org.saltyrtc.client.messages.ResponderServerAuth;
+import org.saltyrtc.client.messages.s2c.NewInitiator;
+import org.saltyrtc.client.messages.s2c.NewResponder;
+import org.saltyrtc.client.messages.s2c.ResponderServerAuth;
 import org.saltyrtc.client.messages.Restart;
-import org.saltyrtc.client.messages.SendError;
-import org.saltyrtc.client.messages.ServerHello;
-import org.saltyrtc.client.messages.Token;
+import org.saltyrtc.client.messages.s2c.SendError;
+import org.saltyrtc.client.messages.s2c.ServerHello;
+import org.saltyrtc.client.messages.c2c.Token;
 
 import java.util.HashMap;
 import java.util.Map;
